Bug#17700: mysql_fix_privilege_tables cannot handle a password \ 	with embedded spaces
Problem:  Original code written assuming password contains no spaces.
Further, because of how sh interprets characters in variables, you
cannot do the "simple solution".
  
Solution:  Move the "was a password provided" check to the end. This
allows us to correctly quote the password argument when passed to the
mysql client program.  Addendum:  Put this check inside of a shell fn
to simplify the code.