Bug #24487 Valgrind: uninited byte in table->record[1] in binlog code for rbr + inno db
The reason of this valgrind's compaint is not a bug but rather a feature of bitwise ops:
for any value of the byte x
x | 1 -> 1,  and x & 0 -> 0.
x, being a null_byte part of record[1] can be left unassigned even after
ha_innobase::index_read_idx because the above and still be correct.
Addding a check memory upon the invocation of the function can detect this fact
long before record[1], old record, is eventually passed to my_write.

Fixed with initialization of record[1]'s null_bytes part in open_table_from_share.


sql/table.cc:
  initialization of null-bytes for valgrind.