Change ::fileno() to fileno() in ndb/common/util/File.cpp.
fileno(), and many other stdio functions, may be macros
(at least on *BSD).

Don't explicitly qualify it with global scope (::fileno()),
because it causes syntax errors.

<cstdio> doesn't define *most* of these macros.

It still has a macro for fileno(), however, on FreeBSD 6.1,
so even using <cstdio> isn't a fix for this.  Better to
just avoid ::standard_library_function() in general, in
order to be portable.


storage/ndb/src/common/util/File.cpp:
  fileno(), and many other stdio functions, may be macros
  (at least on *BSD).
  
  Don't explicitly qualify it with global scope (::fileno()),
  because it causes syntax errors.
  
  Note that <cstdio> doesn't define *most* of these macros.
  It still has a macro for fileno(), however, on FreeBSD 6.1,
  so even using <cstdio> isn't a fix for this.  Better to
  just avoid ::standard_library_function() in general, in
  order to be portable.