BUG#21132 (Slave fails to reconnect on update_slave_list):
The update_slave_list() call is a remainder from attempts to implement failsafe 
replication. This code is now obsolete and not maintained (see comments in 
rpl_failsafe.cc). 

Inspecting the code one can see that this function do not interferre with normal 
slave operation and thus can be safely removed. This will solve the issue 
reported in the bug (errors on slave reconnection). 

A related issue is to remove unneccessary reconnections done by slave. This is 
handled in the patch for BUG#20435.


sql/slave.cc:
  Removed call to obsolete function update_slave_list() upon slave's I/O thread 
  reconnect.