Bug #24998  	rpl_row_delayed_ins.test fails in pushbuild
The test uses show binlog event which is not deterministic due to the single insert delayed
query can generate up to number of inserted rows row-events pair (table_map + Write_row)

The solution is to leave the current binlogging behaviour as it is and change 
the test as spliting arguments of insert delayed query. Note, that such fix was applied
earlier for  binlog_insert_delayed.test :
https://intranet.mysql.com/secure/apps/irclog.php?channel=22&start_time=2006-09-27 

There are no tests with insert delayed and show binlog events combination requiring
this fix.