branches/zip: Do not decompress blocks in buf_page_init_for_read(), but in buf_page_get_gen().  Adjust comments.  This prevents unnecessary decompression in read-ahead when the compressed block exists in the buffer pool.
If the block does not exist in the buffer pool, read-ahead will still
allocate an uncompressed page and decompress the block.

Move buf_block_init_low() and buf_zip_decompress() earlier in the file,
because some compilers are unable to inline functions that are defined
after the invocation.

buf_page_get_gen(): Decompress the block if needed.

buf_page_init_for_read(): Do not decompress.