Bug#25460 - High concurrency MyISAM access causes severe mysqld crash.
Under high load it was possible that memory mapping was started on a table
while other threads were working with the table.

I fixed the start of memory mapping so that it is done at the first table
open or when the requesting thread is using the table exclusively only.