Bug#26782 - Patch: myisamchk -rq creates .MYI.MYI file             on packed MyISAM tables
When fixing the indexes with "myisamchk -rq" after compressing
the table with "myisampack", an optionally supplied extension
".MYI" of the index file was not detected. The extension was
appended unconditionally. The result was ".MYI.MYI".

Now an extension is no longer appended if present already.
Thanks to David Shrewsbury for providing this patch.

Another problem was a misplaced parenthesis. We did never unpack
the file name ("~/..") and always returned a real path.

No test case. This is manually tested with the utilities
"myisampack" and "myisamchk".