BUG#26952: mysql.server needs to be able to not timeout in certain situations
For systems running MySQL through heartbeat, it is imperitive that the startup scripts 
not only return correct return values, but do not return until success or failure has been 
determined. This is a different behavior than is typically wanted for the startup 
of a normal machine. 

This patch adds support for a timeout variable for mysql.server. Read from my.cnf, this
variable defaults to 900 (the current default). A value of 0 means not to wait at all for
startup confirmation. A negative value means to wait forever.